+++ rc/build.h
@@ -2,6 +2,6 @@
 #define STRINGIZE2(s) #s
 #define STRINGIZE(s) STRINGIZE2(s)
 #define VERSION_MAJOR            5
-#define VERSION_MINOR            6
-#define VERSION_REVISION         6
-#define VERSION_BUILD            2014
+#define VERSION_MINOR            7
+#define VERSION_REVISION         0
+#define VERSION_BUILD            2033